requirements.txt
coreapi==2.3.3
Django==1.11
django-debug-toolbar==1.8
django-extensions==1.9.1
django-model-utils==3.1.1
django-grappelli==2.10.1
djangorestframework==3.7.7
dj-database-url==0.4.2
Faker===0.8.12
flake8==3.3.0
ipdb==0.10.3
isort==4.2.5
psycopg2==2.7.1
pytest==3.2.3
pytest-django==3.1.2
uWSGI==2.0.15
